refactor: move domain schemas to Catalog module
